backend/django_api/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
transformers>=4.35.0
torch>=2.1.0
spacy>=3.7.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.1.0
pymongo>=4.6.0
celery>=5.3.0
redis>=5.0.0
python-multipart>=0.0.6
fastapi>=0.100.0
uvicorn>=0.20.0 